requirements.txt
tensorflow==2.9.0
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api
